py/requirements.txt
pandas==2.2.2
mysql-connector-python>=8.0.0
openpyxl>=3.1.2
xlrd==2.0.1